+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+# V0.2
+- 实现了GET，PUT，DELETE请求，并通过了相关测试`/tests`。
+- Able to run GET, PUT, DELETE request, associated tests in the directory `/test`.
+
 # V0.1
 - 实现基础的GET请求，并通过`getAllFeedStatus`方法进行了测试
 - Able to run GET request, tested via `getAllFeedStatus` method